+++ scripts/constants.js
@@ -1,0 +1,3 @@
+export const MODULE_ID = "macro-manager";
+export const TEMPLATE_BUILDER = `modules/${MODULE_ID}/templates/builder.hbs`;
+export const TEMPLATE_WINDOW = `modules/${MODULE_ID}/templates/window.hbs`;